+++ src/main/java/com/example/todo/controller/MvcController.java
@@ -1,12 +1,12 @@
-package com.example.todo.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MvcController {
-    @GetMapping
-    public String toDoMainPage (){
-        return "index";
-    }
-}
+//package com.example.todo.controller;
+//
+//import org.springframework.stereotype.Controller;
+//import org.springframework.web.bind.annotation.GetMapping;
+//
+//@Controller
+//public class MvcController {
+//    @GetMapping
+//    public String toDoMainPage (){
+//        return "index";
+//    }
+//}
