+++ src/main/java/com/example/todo/model/Todo.java
@@ -1,4 +1,29 @@
 package com.example.todo.model;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.Date;
+
+@Entity
+@Data
+@NoArgsConstructor
 public class Todo {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+    private String task;
+    private Date deadline;
+    private Boolean status;
+
+    public Todo(String task, Date deadline, Boolean status) {
+        this.task = task;
+        this.deadline = deadline;
+        this.status = status;
+    }
 }
